refactor: clean-up adhoc single component.
- Remove the unneeded header.
- Move file upload callback to its own util to be reused.
- Embed the file uploader in the flamebeearer.
- Remove unused stuff.
