+++ webapp/javascript/components/AdhocSingle.jsx
@@ -1,11 +1,9 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
-import Dropzone from 'react-dropzone';
 import 'react-dom';
 
 import { bindActionCreators } from 'redux';
 import FlameGraphRenderer from './FlameGraph';
-import Header from './Header';
 import Footer from './Footer';
 import { buildRenderURL } from '../util/updateRequests';
 import {
@@ -13,45 +11,26 @@ import {
   fetchPyrescopeAppData,
   abortTimelineRequest,
 } from '../redux/actions';
-import FileUploader from './FileUploader';
-import { deltaDiffWrapper } from '../util/flamebearer';
+import onFileUpload from '../util/onFileUpload';
 
 function AdhocSingle(props) {
-  const { actions, renderURL, single } = props;
-  const prevPropsRef = useRef();
+  const { actions, renderURL } = props;
   const [flamebearer, setFlamebearer] = useState();
 
   useEffect(() => {
-    if (prevPropsRef.renderURL !== renderURL) {
-      actions.fetchPyrescopeAppData(renderURL);
-    }
-
     return actions.abortTimelineRequest;
   }, [renderURL]);
 
-  const onFileUpload = (data) => {
-    if (!data) {
-      setFlamebearer(null);
-      return;
-    }
-
-    const { flamebearer } = data;
-
-    const calculatedLevels = deltaDiffWrapper(
-      flamebearer.format,
-      flamebearer.levels
-    );
-
-    flamebearer.levels = calculatedLevels;
-    setFlamebearer(flamebearer);
-  };
+  const onUpload = (data) => onFileUpload(data, setFlamebearer);
 
   return (
     <div className="pyroscope-app">
       <div className="main-wrapper">
-        <Header />
-        <FileUploader onUpload={onFileUpload} />
-        <FlameGraphRenderer flamebearer={flamebearer} viewType="single" />
+        <FlameGraphRenderer
+          flamebearer={flamebearer}
+          uploader={onUpload}
+          viewType="single"
+        />
       </div>
       <Footer />
     </div>
